+++ packages/actions/src/morph.ts
@@ -9,13 +9,16 @@ import {
   parseHTMLFragment,
   isElementOrText,
   isElement,
+  focusNextElement,
+  type FocusNextOptions,
 } from '@coldwired/utils';
 
 import { Metadata } from './metadata';
 
-type MorphOptions = {
+type MorphOptions = FocusNextOptions & {
   childrenOnly?: boolean;
   forceAttribute?: string;
+  permanentAttribute?: string;
   metadata?: Metadata;
 };
 
@@ -79,6 +82,7 @@ function morphToDocumentFragment(
 
 function morphToElement(fromElement: Element, toElement: Element, options?: MorphOptions): void {
   const forceAttribute = options?.forceAttribute;
+  const permanentAttribute = options?.permanentAttribute;
 
   morphdom(fromElement, toElement, {
     childrenOnly: options?.childrenOnly,
@@ -94,6 +98,13 @@ function morphToElement(fromElement: Element, toElement: Element, options?: Morp
 
       if (fromElement.isEqualNode(toElement)) {
         return false;
+      }
+
+      if (!force && permanentAttribute) {
+        const permanent = !!fromElement.closest(`[${permanentAttribute}]`);
+        if (permanent) {
+          return false;
+        }
       }
 
       if (!force && metadata) {
@@ -122,6 +133,12 @@ function morphToElement(fromElement: Element, toElement: Element, options?: Morp
         }
       }
 
+      return true;
+    },
+    onBeforeNodeDiscarded(node) {
+      if (isElement(node)) {
+        focusNextElement(node, options);
+      }
       return true;
     },
   });