+++ packages/actions/src/schema.ts
@@ -1,11 +1,15 @@
 export type Schema = {
   forceAttribute: string;
-  focusAttribute: string;
+  focusGroupAttribute: string;
+  focusDirectionAttribute: string;
+  permanentAttribute: string;
   hiddenClassName: string;
 };
 
 export const defaultSchema: Schema = {
   forceAttribute: 'data-turbo-force',
-  focusAttribute: 'data-turbo-focus',
+  focusGroupAttribute: 'data-turbo-focus-group',
+  focusDirectionAttribute: 'data-turbo-focus-direction',
+  permanentAttribute: 'data-turbo-permanent',
   hiddenClassName: 'hidden',
 };