+++ packages/turbo-stream/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @coldwired/turbo-stream
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @coldwired/actions@0.10.0
+  - @coldwired/utils@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes